+++ src/ErrorState.cs
@@ -18,7 +18,7 @@ public readonly partial struct ErrorState
         _isError = isFail;
         _error = error;
     }
-    public ErrorState(ErrorState other) 
+    public ErrorState(ErrorState other)
         : this(other._isError, other._error) { }
 }
 
@@ -77,4 +77,40 @@ partial struct ErrorState
 
     public static ErrorState<TError> Success<TError>() => new();
     public static ErrorState<TError> Error<TError>(TError error) => new(true, error);
+
+    public static ErrorState CombineErrors(ErrorState a, ErrorState b) => (a._isError, b._isError) switch
+    {
+        (false, false) => Success(),
+        (true, false) => a._error,
+        (false, true) => b._error,
+        (true, true) => new AggregateException(a._error, b._error),
+    };
+
+    public static ErrorState<E> CombineErrors<E>(ErrorState<E> a, ErrorState<E> b, Func<E, E, E> errorCombiner) => (a._isError, b._isError) switch
+    {
+        (false, false) => Success<E>(),
+        (true, false) => a._error,
+        (false, true) => b._error,
+        (true, true) => errorCombiner(a._error, b._error),
+    };
+
+    public static ErrorState<E> CombineErrors<E, TArg>(ErrorState<E> a, ErrorState<E> b, TArg arg, Func<E, E, TArg, E> errorCombiner) where TArg : allows ref struct => (a._isError, b._isError) switch
+    {
+        (false, false) => Success<E>(),
+        (true, false) => a._error,
+        (false, true) => b._error,
+        (true, true) => errorCombiner(a._error, b._error, arg),
+    };
+
+    public static ErrorState CombineErrors(ErrorState a, ErrorState b, ErrorState c) => (a._isError, b._isError, c._isError) switch
+    {
+        (false, false, false) => Success(),
+        (true, false, false) => a._error,
+        (false, true, false) => b._error,
+        (false, false, true) => c._error,
+        (true, true, false) => new AggregateException(a._error, b._error),
+        (true, false, true) => new AggregateException(a._error, c._error),
+        (false, true, true) => new AggregateException(b._error, c._error),
+        (true, true, true) => new AggregateException(a._error, b._error, c._error),
+    };
 }